# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "remo-code hub",
    "version": "0.1.0",
    "description": "REST API for the remo-code hub. NOTE: Only the sample `/api/profile/cost-today` route is currently in the spec; the rest of the hub is plain Hono and will be migrated incrementally."
  },
  "servers": [
    {
      "url": "https://app.remo-code.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3040",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {},
    "parameters": {}
  },
  "paths": {
    "/api/profile/cost-today": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Today's spend + daily cost cap",
        "description": "Returns the authenticated user's scheduled-task spend so far today, their configured daily cap, and percent consumed. Used by the cost-cap UI banner.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cost snapshot for the current calendar day in the user's timezone",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cost_usd": {
                      "type": "number"
                    },
                    "cap_usd": {
                      "type": "number"
                    },
                    "percent": {
                      "type": "number"
                    },
                    "timezone": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "cost_usd",
                    "cap_usd",
                    "percent",
                    "timezone"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}